refactor: star 컴포넌트 분리
